move tracking to _document level
