+++ components/layouts/main.tsx
@@ -1,7 +1,5 @@
 import { useState, useEffect, Children, cloneElement } from "react";
 import Head from "next/head";
-import { GoogleAnalytics } from "../googleAnalytics";
-import AdConversionScript from "../adConversionsScript";
 import Navigation from "../navigation";
 import Footer from "../footer";
 import { Alert } from "react-bootstrap";
@@ -165,9 +163,6 @@ function PageTemplate({ children }) {
         )}
         <Footer />
       </div>
-
-      <GoogleAnalytics />
-      <AdConversionScript />
     </>
   );
 }
